+++ swagger-request-validator-core/src/test/java/com/atlassian/oai/validator/interaction/ApiOperationResolverTest.java
@@ -42,7 +42,7 @@ public class ApiOperationResolverTest {
 
         final SwaggerParseResult swaggerParseResult = new OpenAPIParser().readLocation(FILENAME_API_WITH_POST, null, parseOptions);
         final OpenAPI api = swaggerParseResult.getOpenAPI();
-        classUnderTest = new ApiOperationResolver(api, null);
+        classUnderTest = new ApiOperationResolver(api, null, true);
     }
 
     @Parameters(name = "{0}")
@@ -69,6 +69,8 @@ public class ApiOperationResolverTest {
 
                 {"doesNotMatch_whenNoPathMatches", GET, "/not/a/match", missingPath()},
                 {"doesNotMatch_whenNoPathMatches_whenSimilarToActualPath", POST, "/updates/{id}/{action}", missingPath()},
+                {"doesNotMatch_whenTrailingSlashOnRequest", GET, "/path/without/trailing/slash/", missingPath()},
+                {"doesNotMatch_whenTrailingSlashMissingOnRequest", GET, "/path/with/trailing/slash", missingPath()},
 
                 {"doesNotMatch_whenMethodNotAllowed", DELETE, "/id", operationNotAllowed()},
                 {"doesNotMatch_whenMethodNotAllowed_multiplePathParams", GET, "/update/id/action", operationNotAllowed()},